bingo_v2/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
Pillow==10.0.0
Werkzeug==2.3.7
gunicorn==21.2.0